backend/requirements.txt
livekit-agents==0.8.0
google-generativeai==0.3.2
python-dotenv==1.0.0
sentence-transformers==2.2.2
numpy==1.24.3
json5==0.9.14
asyncio